requirements.txt
torch 
torchvision 
torchaudio
Pillow
opencv-python==4.5.1.48
tqdm
